+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ### 0.2.2-pre
 * Add result timestamp (when check is started)
 * Rename http(s) attributes: statusCodes -> statusCode, texts -> text
-* All checker attributes support <attribute> and <attribute>-or
+* All checker attributes handle <attribute> and <attribute>-or
 * Add health#clearCache
-* All checker attributes support regular expression
+* All checker attributes handle regular expression
+* Add threshold support to resource group
 
 ### 0.2.1
 * Add file checker